+++ lib/climb/climb_view.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sendrax/models/attempt.dart';
 import 'package:sendrax/models/climb.dart';
+import 'package:sendrax/models/location.dart';
 import 'package:sendrax/navigation_helper.dart';
 import 'package:sendrax/util/constants.dart';
 
@@ -10,10 +11,16 @@ import 'climb_bloc.dart';
 import 'climb_state.dart';
 
 class ClimbScreen extends StatefulWidget {
-  ClimbScreen({Key key, @required this.climb, @required this.sections, @required this.categories})
+  ClimbScreen(
+      {Key key,
+      @required this.climb,
+      @required this.selectedLocation,
+      @required this.sections,
+      @required this.categories})
       : super(key: key);
 
   final Climb climb;
+  final SelectedLocation selectedLocation;
   final List<String> sections;
   final List<String> categories;
 
@@ -49,7 +56,8 @@ class ClimbWidget extends StatelessWidget {
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.edit),
-            onPressed: () => _editClimb(widget.climb, widget.sections, widget.categories),
+            onPressed: () => _editClimb(
+                widget.climb, widget.sections, widget.categories, widget.selectedLocation),
           )
         ],
       ),
@@ -267,8 +275,10 @@ class ClimbWidget extends StatelessWidget {
         ));
   }
 
-  void _editClimb(Climb climb, List<String> sections, List<String> categories) {
-    NavigationHelper.navigateToCreateClimb(widgetState.context, climb, sections, categories, true,
+  void _editClimb(Climb climb, List<String> sections, List<String> categories,
+      SelectedLocation selectedLocation) {
+    NavigationHelper.navigateToCreateClimb(
+        widgetState.context, climb, selectedLocation, sections, categories, true,
         addToBackStack: true);
   }